Small water bug fixed
Had some problems with changing the players velocity. Solved this by creating a seperate function almost that determines if the play is in or out of the water (Using my tiles) and from that, I was able to change the velocity. I believe that it is 5 without water and 2 with water (Pixels).