Collision with blocks | Hitboxes
Changed the Hit-box of the player. Before it was a rect but that made it so that I couldn't detect which side the player was coliding with. Instead I made 4 different rectangles that are all 1 pixel wide and or high, and it is able to detect whether that collides.

Works well as of right now, I am going to set some Boolian values to the movement, and only allow them to be true when the player is not coliding with something.