Working collisions | Hitbox update
Added collisions. Don't know if it was the most effective way but it works. Basically if you collide with a wall, then that movement is false. I had an issue with the hitbox, because it is the shape of the box, it was basically procing three walls when one collided (E.G You hit the top wall, the left and right side would collide because they were the same size). This was resolved by changing the size of the hitbox. They are basically just lines but I use the Rect argument due to the collide ability. Game working well as of now, no known errors. Going to try and add enemies now I reckon.