Clothing
Just added a pretty basic clothing menu. Is able to display the user's outfit. Am going to now change it in the while loop. Should be easy enough.

Only problem I had was with the following:

```
                if imgnumb_1 == -5:
                    imgnumb_1 = 0
                elif imgnumb_1 == 4:
                    imgnumb_1 = -1
```
I didn't know how to address the negative numbers, but I was able to complete it in the end.